Modification by Gen LI Run the reformat code and fix weak warnings.
