+++ src/main/java/message/Message_R.java
@@ -5,8 +5,10 @@ import order.Order;
 
 import java.util.Date;
 import java.util.HashMap;
+
 /**
  * This class is an entity class of the popup message sent to customer or admin.
+ *
  * @author GEN LI (Reagan)
  */
 
@@ -26,7 +28,7 @@ public class Message_R {
      */
 
 
-    public Message_R(Order order){
+    public Message_R(Order order) {
         this.item_list = order.get_item_list();
         this.order_date = order.get_date();
         this.id = order.get_id();
@@ -38,46 +40,51 @@ public class Message_R {
 
     /**
      * This is a method that is used to get the item_list stored in the order entity.
+     *
      * @return the method returns a HashMap of items required in the order.
      */
 
-    public HashMap<Item, Integer> get_item_list(){
+    public HashMap<Item, Integer> get_item_list() {
         return this.item_list;
     }
 
     /**
      * This is a method that is used to get the date of order stored in the order entity.
+     *
      * @return the method returns a Date type variable of date of the order.
      */
 
-    public Date get_order_date(){
+    public Date get_order_date() {
         return this.order_date;
     }
 
     /**
      * This is a method that is used to get the order id of the order.
+     *
      * @return the method returns a String of order_id.
      */
 
-    public String get_order_id(){
+    public String get_order_id() {
         return this.id;
     }
 
     /**
      * This is a method that is used to get the status of the order.
+     *
      * @return the method returns a String of status of the order.
      */
 
-    public String get_status(){
+    public String get_status() {
         return this.status;
     }
 
     /**
      * This is a method that is used to get the generated date of the order.
+     *
      * @return the method returns a Date type variable of order's date of generating.
      */
 
-    public Date get_message_date(){
+    public Date get_message_date() {
         return this.message_date;
     }
 }
